messages-grip-bot/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.0
asyncpg==0.30.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
exceptiongroup==1.2.2
fastapi==0.115.4
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pyaes==1.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pyrogram @ https://github.com/KurimuzonAkuma/pyrogram/archive/dev.zip
PySocks==1.7.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
TgCrypto==1.2.5
typing_extensions==4.12.2
uvicorn==0.32.0
